Moves reuse_functions to experimental
