+++ hamilton/function_modifiers/__init__.py
@@ -1,6 +1,6 @@
 import logging
 
-from . import base, configuration, dependencies, expanders, macros, metadata, reuse, validation
+from . import base, configuration, dependencies, expanders, macros, metadata, validation
 
 logger = logging.getLogger(__name__)
 
@@ -58,7 +58,3 @@ check_output = validation.check_output
 check_output_custom = validation.check_output_custom
 IS_DATA_VALIDATOR_TAG = validation.IS_DATA_VALIDATOR_TAG
 DATA_VALIDATOR_ORIGINAL_OUTPUT_TAG = validation.DATA_VALIDATOR_ORIGINAL_OUTPUT_TAG
-
-# reuse-subdag
-
-reuse_functions = reuse.reuse_functions
